refactor(#17): move tags list component to shared dir
