+++ src/app/App/MainScreen/NotesOverview/index.tsx
@@ -6,9 +6,9 @@ import { cn } from '@bem-react/classname';
 import { ITag } from '../../../../core/Registry/Tags/Tags';
 import { $activeTag, $tags, setActiveTag, tagsChanged } from '../../../../core/state/tags';
 import { Icon } from '../../../components/Icon/Icon.bundle/common';
+import { List } from '../../../components/List';
 import { useTagsRegistry } from '../../Providers';
 
-import { List } from './List';
 import { TagEditor, TagEditorData } from './TagEditor';
 import { TagItem, TagsList } from './TagsList';
